Add a shell intrinsic to the type lanugage
Similar to `Uppercase<S>` and `Lowercase<s>` and so on, add a type
`Shell<S extends string>` which runs the string S as a
`child_process`, captures its stdout, and produces the string literal
type that is the captured output string.

For example, `Shell<"echo hello">` is the type `"hello"`.